A Helm chart for a cloudflare tunnel.
This is in reference to github issue #16. I saw the example yaml in
the named-tunnel-k8s directory and thought it would be much more
useful as a helm chart. And I had to make a helm chart anyway so I
could deploy it with Terraform.  Here's the result, and figured I'd
contribute it back.

Not sure how you'd like to work this into examples or have it be
stand-alone, so here's a start in a stand-alone directory.